+++ src/components/bluepad32/uni_hid_parser_ds4.c
@@ -397,9 +397,9 @@ static void ds4_parse_input_report_01(uni_hid_device_t* d, const ds4_input_repor
     if (r->buttons[1] & 0x08)
         ctl->gamepad.buttons |= BUTTON_TRIGGER_R;  // R2
     if (r->buttons[1] & 0x10)
-        ctl->gamepad.misc_buttons |= MISC_BUTTON_BACK;  // Share
+        ctl->gamepad.misc_buttons |= MISC_BUTTON_SELECT;  // Share
     if (r->buttons[1] & 0x20)
-        ctl->gamepad.misc_buttons |= MISC_BUTTON_HOME;  // Options
+        ctl->gamepad.misc_buttons |= MISC_BUTTON_START;  // Options
     if (r->buttons[1] & 0x40)
         ctl->gamepad.buttons |= BUTTON_THUMB_L;  // Thumb L
     if (r->buttons[1] & 0x80)
@@ -447,9 +447,9 @@ static void ds4_parse_input_report_11(uni_hid_device_t* d, const ds4_input_repor
     if (r->buttons[1] & 0x08)
         ctl->gamepad.buttons |= BUTTON_TRIGGER_R;  // R2
     if (r->buttons[1] & 0x10)
-        ctl->gamepad.misc_buttons |= MISC_BUTTON_BACK;  // Share
+        ctl->gamepad.misc_buttons |= MISC_BUTTON_SELECT;  // Share
     if (r->buttons[1] & 0x20)
-        ctl->gamepad.misc_buttons |= MISC_BUTTON_HOME;  // Options
+        ctl->gamepad.misc_buttons |= MISC_BUTTON_START;  // Options
     if (r->buttons[1] & 0x40)
         ctl->gamepad.buttons |= BUTTON_THUMB_L;  // Thumb L
     if (r->buttons[1] & 0x80)