+++ src/main/sketch.cpp
@@ -158,21 +158,27 @@ void loop() {
             // miscButtons() which return a bitmask.
             // Some gamepads also have DPAD, axis and more.
             Console.printf(
-                "idx=%d, dpad: 0x%02x, buttons: 0x%04x, axis L: %4d, %4d, axis R: %4d, "
-                "%4d, brake: %4d, throttle: %4d, misc: 0x%02x\n",
-                i,                        // Gamepad Index
-                myGamepad->dpad(),        // DPAD
-                myGamepad->buttons(),     // bitmask of pressed buttons
-                myGamepad->axisX(),       // (-511 - 512) left X Axis
-                myGamepad->axisY(),       // (-511 - 512) left Y axis
-                myGamepad->axisRX(),      // (-511 - 512) right X axis
-                myGamepad->axisRY(),      // (-511 - 512) right Y axis
-                myGamepad->brake(),       // (0 - 1023): brake button
-                myGamepad->throttle(),    // (0 - 1023): throttle (AKA gas) button
-                myGamepad->miscButtons()  // bitmak of pressed "misc" buttons
+                "idx=%d, dpad: 0x%02x, buttons: 0x%04x, axis L: %4d, %4d, axis R: %4d, %4d, brake: %4d, throttle: %4d, "
+                "misc: 0x%02x, gyro x:%6d y:%6d z:%6d, accel x:%6d y:%6d z:%6d\n",
+                i,                         // Gamepad Index
+                myGamepad->dpad(),         // DPAD
+                myGamepad->buttons(),      // bitmask of pressed buttons
+                myGamepad->axisX(),        // (-511 - 512) left X Axis
+                myGamepad->axisY(),        // (-511 - 512) left Y axis
+                myGamepad->axisRX(),       // (-511 - 512) right X axis
+                myGamepad->axisRY(),       // (-511 - 512) right Y axis
+                myGamepad->brake(),        // (0 - 1023): brake button
+                myGamepad->throttle(),     // (0 - 1023): throttle (AKA gas) button
+                myGamepad->miscButtons(),  // bitmak of pressed "misc" buttons
+                myGamepad->gyroX(),        // Gyro X
+                myGamepad->gyroY(),        // Gyro Y
+                myGamepad->gyroZ(),        // Gyro Z
+                myGamepad->accelX(),       // Accelerometer X
+                myGamepad->accelY(),       // Accelerometer Y
+                myGamepad->accelZ()        // Accelerometer Z
             );
 
-            // You can query the axis and other properties as well. See Gamepad.h
+            // You can query the axis and other properties as well. See ArduinoController.h
             // For all the available functions.
         }
     }